feat: 🎸 add react@18 support
in order to support react 17 and older the mount/unmount logic is placed
behind a adapter. If an application needs to be exported with an older
react-version it can use @navikt/navspa-react-17 or provide its own
adapter

BREAKING CHANGE: 🧨 switched to react@18. Apps that export itself and can not be updated to
react@18 needs to add the react-17 adapter. Consumer-apps needs no
changes